Add files via upload
cheduling of real time periodic and aperiodic tasks with event handling and setting specific priorities for tasks in the linux environment .
This involved the concepts of POSIX threads in Linux environment .Involved creation, running, suspending and killing of multiple pthreads. Gained good understanding of the thread lifecycle.
Building , cross compiling and setting up the software development environment on the target board of Intel Galileo Gen 2 to run the user application.
Analysis of CPU time scheduling with Linux Trace tools of Kernelshark.